make mystock

+++ catalog.py
@@ -110,8 +110,11 @@ class XmlGetter(Resource):
                 continue
             all_count +=1
             if row['id'] not in codes:
-                couch.deleteDoc(row['id'], row['value'])
-                deleted_count +=1
+                if 'mystock' in row['value'] and row['value']['mystock']>0:
+                    pass
+                else:
+                    couch.deleteDoc(row['id'], row['value'])
+                    deleted_count +=1
         # destroy cache here!
         from pc import root
         root.clear_cache()
